+++ packages/core/lib/core.js
@@ -8,8 +8,16 @@ import {
 import { Input } from "@wpmudev/react-input";
 import { Button } from "@wpmudev/react-button";
 import { Modal } from "@wpmudev/react-modal";
+import { ButtonIcon } from "@wpmudev/react-button-icon";
 import { Post } from "@wpmudev/react-post";
 import { Notifications } from "@wpmudev/react-notifications";
+import { Dropdown } from "@wpmudev/react-dropdown";
+import {
+	Accordion,
+	AccordionItem,
+	AccordionItemHeader,
+	AccordionItemBody
+} from "@wpmudev/react-accordion";
 
 export {
 	Box,
@@ -20,6 +28,12 @@ export {
 	Input,
 	Button,
 	Modal,
+	ButtonIcon,
 	Post,
-	Notifications
+	Notifications,
+	Dropdown,
+	Accordion,
+	AccordionItem,
+	AccordionItemHeader,
+	AccordionItemBody
 };