+++ packages/react-button-icon/lib/react-button-icon.js
@@ -1,6 +1,13 @@
 import React from "react";
 
-const ButtonIcon = ({ label, icon, iconSize, extraClasses, color, design = "solid", ...props }) => {
+const ButtonIcon = React.forwardRef( ({
+	label,
+	icon,
+	design = "solid",
+	color,
+	className,
+	...props
+}, ref) => {
 	const loader = (
 		<span
 			className="sui-icon-loader sui-loading"
@@ -16,7 +23,9 @@ const ButtonIcon = ({ label, icon, iconSize, extraClasses, color, design = "soli
 		</React.Fragment>
 	);
 
-	let className = `sui-button-icon ${extraClasses || ""}`;
+	className = '' !== className
+		? 'sui-button-icon ' + className
+		: 'sui-button-icon';
 
 	// Set button color.
 	switch (color) {
@@ -57,6 +66,7 @@ const ButtonIcon = ({ label, icon, iconSize, extraClasses, color, design = "soli
 	if (props.href) {
 		return (
 			<a
+				ref={ref}
 				className={className}
 				disabled={props.disabled || props.loading}
 				{...props}>
@@ -67,12 +77,13 @@ const ButtonIcon = ({ label, icon, iconSize, extraClasses, color, design = "soli
 
 	return (
 		<button
+			ref={ref}
 			className={className}
 			disabled={props.disabled || props.loading}
 			{...props}>
 			{props.loading ? loader : content}
 		</button>
 	);
-};
+});
 
 export { ButtonIcon };