+++ packages/react-button/lib/button.js
@@ -1,6 +1,13 @@
 import React from "react";
 
-const Button = ({ label, icon, design = "solid", color, ...props }) => {
+const Button = ({
+	label,
+	icon,
+	design = "solid",
+	color,
+	className,
+	...props
+}) => {
 	const loader = (
 		<span
 			className="sui-icon-loader sui-loading"
@@ -18,7 +25,7 @@ const Button = ({ label, icon, design = "solid", color, ...props }) => {
 		</React.Fragment>
 	);
 
-	let className = "sui-button";
+	className = `sui-button${ className ? ' ' + className : '' }`;
 
 	// Set button color.
 	switch (color) {
@@ -56,25 +63,22 @@ const Button = ({ label, icon, design = "solid", color, ...props }) => {
 		className += " sui-button-onload";
 	}
 
-	if (props.href) {
-		return (
-			<a
-				className={className}
-				disabled={props.disabled || props.loading}
-				{...props}>
-				{props.loading ? loader : content}
-			</a>
-		);
+	let htmlTag = 'button';
+	if ( props.href ) {
+		htmlTag = 'a';
+	} else if ( props.htmlFor ) {
+		htmlTag = 'label';
 	}
 
-	return (
-		<button
-			className={className}
-			disabled={props.disabled || props.loading}
-			{...props}>
-			{props.loading ? loader : content}
-		</button>
-	);
+	return React.createElement(
+		htmlTag,
+		{
+			className: className,
+			disabled: props.disabled || props.loading,
+			...props
+		},
+		props.loading ? loader : content
+	)
 };
 
 export { Button };