+++ packages/react-input/docs/react-input.stories.js
@@ -10,7 +10,6 @@ export const primary = ( args ) => <Input { ...args } />;
 primary.storyName = 'Text Input';
 primary.args = {
     label: 'Company Name',
-    maxLength: '',
     placeholder: 'Eg. Acme Industries',
     description: 'Maecenas sed diam eget risus varius blandit sit amet non magna.'
 };
@@ -36,12 +35,13 @@ primary.argTypes = {
 
 export const secondary = ( args ) => {
 	return (
-		<Input type="number" { ...args } />
+		<Input { ...args } />
 	);
 };
 secondary.storyName = 'Number Input';
 secondary.args = {
-    label: 'Year',
+	type: 'number',
+    label: 'Yearzz',
     placeholder: 'Eg. 10',
     description: 'Maecenas sed diam eget risus varius blandit sit amet non magna.',
 	min: 0